Split types.go
